Added methods for trains and station to handle capacity
and train creation.

+++ src/station/platform.rs
@@ -2,13 +2,13 @@ use crate::train::TrainType;
 
 pub struct Platform {
     id: u8,
-    station_id: u16, 
+    station_id: u32, 
     occupied: bool,
     platform_type: TrainType,
 }
 
 impl Platform {
-    pub fn new(id: u8, station_id: u16 , platform_type: TrainType) -> Platform {
+    pub fn new(id: u8, station_id: u32 , platform_type: TrainType) -> Platform {
         
         
         Platform {
